+++ packages/frameworks/angular/src/radio/radio-group/index.ts
@@ -1,4 +1,5 @@
 export * from "./radio-group-error-text.component"
+export * from "./radio-group-hint.directive"
 export * from "./radio-group-items.directive"
 export * from "./radio-group-label.directive"
 export * from "./radio-group-root.directive"