+++ packages/common/qds-core/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @qualcomm-ui/qds-core Changelog
 
+## 1.16.0 (2026/01/30)
+
+### Features
+
+- [number-input]: add unit selector
+- [radio]: hint at item level
+
+### Bug Fixes
+
+- [radio]: group hint + CSS fixes
+- [switch]: hint text
+- [checkbox]: hint text
+
+### Miscellaneous Chores
+
+- **deps:** update dependencies [@qualcomm-ui/core]
+
 ## 1.15.1 (2026/01/29)
 
 ### Miscellaneous Chores